chore: merge server and daemon
