+++ cosmic-accounts-daemon/Cargo.toml
@@ -6,10 +6,13 @@ description = "COSMIC Online Accounts daemon service"
 
 [dependencies]
 cosmic-accounts = { path = ".." }
-tokio = { version = "1.0", features = ["full"] }
 zbus = "4.4.0"
-tracing = "0.1"
-tracing-subscriber = "0.3"
+axum = "0.8.4"
+serde = { workspace = true }
+tokio = { workspace = true }
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true }
+
 
 [[bin]]
 name = "cosmic-accounts-daemon"
