--- cosmic-accounts-server/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "cosmic-accounts-server"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-axum = "0.8.4"
-serde = { workspace = true }
-tokio = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
